% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_column_names}
\alias{clean_column_names}
\title{Clean column names}
\usage{
clean_column_names(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame with cleaned column names
}
\description{
Clean column names
}
