% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{compute_avg_amounts}
\alias{compute_avg_amounts}
\title{Compute row-wise averages for BILL_AMT and PAY_AMT columns}
\usage{
compute_avg_amounts(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame with new avg columns and removed original columns
}
\description{
Compute row-wise averages for BILL_AMT and PAY_AMT columns
}
