% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_abstraction.R
\name{evaluate_k_values}
\alias{evaluate_k_values}
\title{Evaluate accuracy of KNN across multiple k values}
\usage{
evaluate_k_values(train, test, train_labels, test_labels, k_values = 1:20)
}
\arguments{
\item{train}{Normalized training predictors}

\item{test}{Normalized test predictors}

\item{train_labels}{Training labels}

\item{test_labels}{Test labels}

\item{k_values}{Vector of k values to evaluate (default: 1 to 20)}
}
\value{
A data frame with k and corresponding accuracy
}
\description{
Evaluate accuracy of KNN across multiple k values
}
