% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_abstraction.R
\name{normalize}
\alias{normalize}
\title{Normalize a numeric vector using Min-Max scaling}
\usage{
normalize(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A normalized vector between 0 and 1
}
\description{
Normalize a numeric vector using Min-Max scaling
}
