% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_abstraction.R
\name{save_accuracy_plot}
\alias{save_accuracy_plot}
\title{Save accuracy vs. k plot to file}
\usage{
save_accuracy_plot(accuracy_df, out_path)
}
\arguments{
\item{accuracy_df}{Data frame with k and accuracy}

\item{out_path}{Output file path for the plot (PNG)}
}
\description{
Save accuracy vs. k plot to file
}
