% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_abstraction.R
\name{save_model_outputs}
\alias{save_model_outputs}
\title{Save confusion matrix and performance metrics}
\usage{
save_model_outputs(predictions, actual, prefix)
}
\arguments{
\item{predictions}{Predicted labels}

\item{actual}{Actual labels}

\item{prefix}{Output file path prefix (no extension)}
}
\description{
Save confusion matrix and performance metrics
}
